Add array casting in additional cases
Reenable the automatic conversion of string value lists to text arrays only
in development and when running frozen tests, and do it for all non-empty
arrays, instead of the default of only arrays of 2 or more elements.

I do not feel comfortable enough to turn this on in production, since it will
not catch places in the tests where we are only using empty arrays, but where
the arrays will be non-empty in production.  I'm not sure we have that case,
but since we don't have the tooling to ensure that, enabling this in production
seems too risky.  Doing it only in the frozen tests is sufficient for now, since
it will still result in the query parameterization log having consistent SQL.
I'm also enabling it in development, because after enough time, if there are any
issues that are not caught by the tests, they should occur in a developer's
development environment, and once reported, they can be easily fixed.

Update the Sequel version to 5.90 as that includes the changes this needs.