Add `Config.is_e2e` to modify prog behavior when running in E2E
While testing postgres HA in E2E tests, we discovered that provisioning
the MinIO cluster fails as it requires distinct storage devices.
Additionally, when `Config.production?` is true, postgres HA would
require separate hosts for the primary and standby. This change
introduces a config flag to relax some of these requirements when
running in the E2E environment.