Fix addressing and subnet size of metadata-endpoint
We were adding a whole prefix for the metadata-endpoint. That is not
needed, we are turning it to a single address. Also, both the bootcmd
and runcmd were being used for non sudo blocker. That was unnecessary,
so we are switching it to only bootcmd so that the rule is added only
once.