Remove unnecessary transactions in progs
We already run the prog labels in a DB transaction when we call
`strand.run`, so there's no need to wrap them in another transaction.

We only use DB transactions in the assemble function in progs because
they might be called from web routes or pry, so they aren't wrapped by a
tra transaction all the time.