Separating slice and no-slice VMs on hosts
The allocator decided if the VM should be created with a slice or without a slice based on the host's accept_slices flag. However that logic only applied when the project had `use_slices_for_allocation` flag set to `true`
When the project flag was false any host could be picked, and a VM with no slice was created. The result was a mix of VMs in-slices and VMs with no-slices on the same host, if the host had `accept_slices=true` flag. On hosts with `accept_slices=false` no slices were created.

This fix adds an extra constraint to not allow placement of no-slice VMs on hosts set with `accept_slices=true` from projects that do not want slices. It also sets a preference to pick a host that accepts slices for projects that do want that.

Tests added to cover all those cases explicitely.