Remove openapi entries using /id/ routes
The routing tree was changed from an "id/" prefix to an "_" prefix for
ubids in the cases were you can provide either a name or an id,
since before the introduction of committee.

Changing the "id/" prefix to "_" breaks committee.  I suppose you could
try to fix that and maybe it makes something nicer, but I'm not sure
what the benefit would be.

You don't actually need these entries, because all of the incorrect
/id/ routes removed are handled by the name routes, because the name
parameters do not specify a format, so the underscored ubid will
match them.