Use the right neg_ttl value while setting up the DnsZone
While setting up the DnsZone, we will create the right config with the
right neg_ttl.

This was needed because during the k8s bootstrap, there was a 4 minute
timeout and if the first dns query failed, there was a 1hour neg_ttl
fo actually retrying the dns query and bootstrap would also fail.