Fix cli pg pg_dumpall command
The pg_dumpall command does not take an argument, it takes the command
line string as part of the -d option. Add special code to bin/ubi to
allow the new argument to appear without --, but it must start with -d.

Change the pg_dumpall command handling to replace the ["--", conn_string]
arguments with ["-d#{conn_string}"].

Split the pg_dumpall spec into its own file, since the behavior is now
different than psql and pg_dump.  Rename the file the psql and pg_dump
to reflect it no longer handles pg_dumpall.

I didn't catch this originally, because it's hard to test PostgreSQL
support in the development environment, due to:

* My development host is in Finland, and we don't support PostgreSQL
  VMs in Finland

* Even if you work around that, you need custom PostgreSQL boot
  images, and I'm not sure how to get access to those in a development
  environment.