Allow cli vm create to support multiple public keys
This worked on the web, but not previously in the cli, and that's
because on the web, newlines are submitted as CRLF, but if you do

  ubi vm loc/name create "$(< ~/.ssh/authorized_keys)"

newlines will be LF, and somehow this breaks something internally
and doesn't allow you to log in after the VM is created.  Converting
the LF to CRLF fixes it.