Hesitate Before Creating a github-runner Capacity Page
We get github-runner capacity pages all the time and usually we do
nothing much but wait for capacity to be available and the page to
auto-resolve. The page severity is low enough to not create a call
to the on-call.

This pattern of accepted ignore of the capacity pages caused us to
miss an actual issue with one of our runner classes. So we agreed
to reduce the sensitivity of our alerting for capacity issues, but
increase the severity so that when it happens, we create a call.