Delete cache entry if it is not committed in 30 mins
Cache entries are saved in 2 steps (i) Reserve (ii) Commit
If an entry is reserved but not committed, it causes subsequent
save requests for that entry to fail. To prevent it removing
non-committed entries after 30 mins. 30 mins is calculated by
adding a 2x buffer after max upload time for an entry with max
size.