Clover part - Retaining bursting limits after VM restart
Vm restart was simply calling systemctl directly and that was wiping out the cpu burst limit from a cgroup settings. Those settings are not persisted anywhere and have to be re-set every time the service restarts. This was handled for the host restart, but not for the VM restart.

Calling the restart script on rhizome added in the previous commit.