Fix 2pc problem between `daemonizer --clean` and hop
Running `daemonizer --clean` has an immediate side effect on the
remote instance, but there's no guarantee that the transaction nesting
the side effect, implied in the label, will commit.  This creates a
problem on retry, as the daemonized process will have been cleaned,
but `prep` would have not recorded it had done so, and then it would
try to run the `prep` daemon again as though nothing ever happened.

Reported-By: Enes Cakir <enes@ubicloud.com>