Adding support for Postgres on Burstables
Allowing Postgres to run in Burstable VMs. This commit is adding a new set of PostgresSize options that defines Postgres VMs running on Burstable-1 and Burstable-2 instances.

In order to differentiate billing counters for PG on Standard vs. PG on Burstable, we are now using a combination of vm_family and pg flavor values to create a resource_family used in billing counters. For the UI the family is for now hardcoded as 'standard' but this will be changed in the next commit.

For small instances the work_mem configuration value for PG instance is now set to a 1MB minimum. This will be used for 2GB and 4GB burstable VMs.