Improve Rodish's error reporting and usage reporting
This make numerous changes to improve error reporting and usage reporting.

This makes the CommandFailure exception class take an array of option
parsers.  It adds CommandFailure#message_with_usage, which can be called
to get the error message and the usage from the option parsers. This
allows failing commands to show the error followed by proper usage.

If there are more than 6 subcommands to a command, this prints the
subcommands each on a separate line after the usage and options.

In cases where you are skipping option parsing, an argument is now
required to set the banner.  This allows for displaying usages for
commands that skip option parsing.  There is a new SkipOptionParser
class for this.

You can now call options and post_options without a block, to only
set the banner/usage without actually accepting options.

Invalid post options are now handled correctly, showing the correct
option parser.

Many error messages have been shortened to remove the display of valid
subcommands, since that information would be redundant now that the
usage (including valid subcommands) can be displayed right after the
error message.

A process_command_failure private method has been added, and both
subcommand and post subcommand processing now use it, showing the
appropriate usage for each on failure.

Command#each_subcommand now yields post subcommands in addition to
normal subcommands.

Command#raise_failure has been added for raising CommandFailure
exceptions with the command's option parsers.

Command#options_text has been added for returning a string with
the usage for all option parsers (or nil if the command has no
option parsers).

Command#subcommand and #post_subcommand have been added, which force
autoloading of the related commands if they are not already loaded.

Command#option_parsers returns an array with the option parser and
post option parser for the command (only including the parsers that
are defined).

Processor#usages now includes usages for post subcommands.

This adds a ProgramBug subclass of CommandFailure, so that consumers
that want to differentiate program bugs from failed commands can easily
do so.