Adding UI support for Burstables
Adding a new selector for `family` to create pages and using it to list VM sizes specific to each family. The visibility of Burstable family, or any other in the future that requires slics support, is guarded by the project feature flag.

Also updated the logic in generate_vm_options to fully use what is set in Options::VmSize and Options::PostgresSize structures. This should make it easier to add more Vm sizes in the future.