Refactor Rodish
Make Rodish::Processor a module, and make Rodish.processor take a
module/class that is extended with Rodish::Processor.
Rodish::Processor#process no longer takes a context keyword
argument, and instead creates a new instance of the class it is
included using the remaining arguments.

This allows for more intuitive behavior.  The before and run
blocks inside the "on" blocks are now in instance scope of the
receiver of the "on" method.