Update model image for 'mistral-small-3'
The model files on Huggingface at
https://huggingface.co/mistralai/Mistral-Small-24B-Instruct-2501/tree/main
basically contain two versions of the model: consolidated.safetensors and
model-*.safetensors. In the previous image we only kept a single version
to save space. In this image we keep both. That allows to switch between
and test both versions.

Run producing the image:
https://github.com/ubicloud/ai-images/actions/runs/13368039715