Pass location while creating new timeline at the end of restore
We used to keep blob storage id in a config variable, but then we started to
search it using location and project id. However, we forgot to pass location
while creating new timeline at the end of restore. This prevents us taking
backups of the restored databases. This commit fixes this issue by passing the
location while creating new timeline at the end of restore.