Calculate VAT for EU countries
We will start charging VAT for EU customers next month.

In this commit, we identify which Ubicloud entity will issue the invoice
and calculate the VAT amount.

I will update the invoice rendering to include VAT information in future
commits.

We will manually enable the `annual_non_dutch_eu_sales_exceed_threshold`
flag when needed.

The above rules are used to determine the issuer and VAT amount for
invoices:

- Invoices are issued by Ubicloud Inc. for non-EU customers without VAT
  applied.
- Invoices are issued by Ubicloud B.V. for EU customers.
    - If the customer has provided a VAT number from the Netherlands, we
      charge 21% VAT.
    - If the customer has provided a VAT number from another European
      country, we include a reverse charge notice along with 0% VAT.
    - If the customer hasn't provided a VAT number, we charge 21% VAT
      until non-Dutch EU sales exceed annual threshold, than we charge
      local VAT.