Refactor cli golden factor specs
Instead of parsing file names and basing the commands on those,
store the commands in files, so that you can edit files to change
which commands are run.  This makes the file names for the golden
files match the command being run, other than replacing / with _.

To add commands for golden files, edit the files in the
spec/routes/api/cli/golden-file-commands folder.