Reduce VM create time
Remove the `wait_for_slice` label from the Vm Nexus and with that remove the wait time before calling the `prep`. The VM service on the host (systemctl service) has an implicit dependency on the slice by specifying the slice name in the service file. That ensures that the VM is not started until the slice is started and ready. That waiting can happen in `wait_for_sshable`, as we are waiting there anyway for the service to start and be connectable. This allows the slice creation to proceed fully parallel to the Vm creation.