Use deadline pages for VM host unavailability
PostgresServer was using deadlines instead of separate unavailability
pages. Hadi recently replaced them with deadline pages for both VM and
VM slice at 57f2d0f7.

The advantage of deadline pages is that they automatically resolve when
the resource is deleted and provide extra context useful for on-call
debugging. Otherwise, operators often forget to resolve them if the
resource is deleted.