Make cli lb list command include location
This requires changes to the api/serializer so that location is
included for load balancers by default.

I think the reason this wasn't included for load balancer but was
included for private subnets is that load balancer does not have
a location column, it gets the location through the private_subnet
it is associated with.