Log runner vcpus instead of cores
With the burstable feature, we set VM cores to 0 and then set the actual
values during provisioning. So, it's 0 for runner allocated logs if it's
provisioned on a VM host with slices enabled. Therefore, we should use
vcpus instead of cores in the logs. I'm keeping the cores since they
have actual values after provisioning, and I don't want to break all the
views in our monitoring tool.