Add cli ps {connect,disconnect} commands
This was not previously supported in the api, so add api support so the cli
can use it.

To keep changes small, this has the api routes match the web routes, so the
connect/disconnect cases are not consistent, as connect takes the id as a
parameter, and disconnect takes the id as a path segment.