Use id instead of ubid in cli usage lines
ubid is an internal concept that users will not be familiar with.
When the cli lists/shows objects, it uses the ubid when displaying
the id, so users already understand that is the id, they never see
the actual uuid.  This should make the usage lines a bit more
understandable.