Introduce new models for LoadBalancerPort & LoadBalancerVmPort
Currently our loadbalancers only support one port. For k8s ingress controllers
we needed to have multiple ports per lb and that was the motivation for this change.

Each LB can have multiple ports as LoadBalancerPort holds a FK to LoadBalancer table

Also in order to implement the healthcheck logic for ecah port, we needed to keep track of
the state of this tuple (lb, port, vm). For that reason, we needed a new table holding FKs
to all of these tables and also an additional state variable.

New logic can be implemented on top of this design.