Add ubid types for 4 models, do not leak TYPE_ETC ubid values
This adds a check in the route specs that TYPE_ETC ubid values are not
present in responses.  It then changes four models to use model-specific
ubid types instead of TYPE_ETC.

The original impetus for this change is to eventually remove
authorization code that explicitly checks for TYPE_ETC and assumes
ApiKey, which will cause problems in the future if we ever have
another model using TYPE_ETC that the authorization system needs
to deal with.