Override Model.ubid_type before including ResourceMethods
This avoids an exception per class during class loading.

For FirewallsPrivateSubnets and CertsLoadBalancers, this defines
a ubid_type method when one was previously not defined.  This
method isn't used, since these models do not have an id field
currently.  I'm guessing these models only use ResourceMethods
for the before_destroy hook, so potentially, that could be moved
to a separate module, or we could use a trigger approach to
updating the archive table instead of a model hook approach.