Fix NoMethodError when viewing VM create page with burstables and location_latitude_fra feature flags enabled
This issue occurred because there is no billing rate for a burstable
size for the latitude_fra location.

Add a BillingRate.unit_price_from_resource_properties that returns nil
instead of raising a NoMethodError if there is no matching resource
property.  Use it to DRY up code in ContentGenerator::{Vm,Postgres}.

Make ContentGenerator::Vm.family not return an array if there are no
valid sizes at the requested location for the family.

This issue affects VMs and not PostgreSQL databases because
Option.postgres_locations does not include the latitude_fra location.