Add libffi to docker image to fix build
Amazingly, the bundling of this dependency was dropped in a point
release update of the Docker image, although the Ruby upstream
approach to libffi changed in the 3.2 series quite a while ago.  While
the changes to image size is appealing, the breakage to builds is less
gratifying.

So, add libffi back in ourselves.