Rhizome: Increase SPDK iobuf sizes.
In ab3933b we increased iobuf sizes to have enough resources for our
current workload. It reduced the error rate across our fleet from 5
times per week to once in every few weeks. This PR further increases the
resources to reduce the chance of error even more.

Since we fixed error handling with a previous release of bdev_ubi-0.3,
the expected behavior in case of exceeding limits is an error instead of
a crash.