Make location references a foreign key, create indexes - Mig 1
This commit adds the location foreign_keys to all of the models that
uses location. This way, we will not be matching the text values but
make use of the database relationship.
We are refilling the location_ids based on the location column of the
entities.

We are also adding a new index on project_id, location_id, name because we
already had an index on project_id, location, name and it will be
dropped once the location column is dropped.