Drop support for optional leading underscore in routes
Also, do not allow `u` as a character, since ubids cannot include it.

This was previously kept for backwards compatibility, but there are no
entries in the request log as far as the logs go back for anyone using
underscored ubids for any of the 5 ubid types.  Since nobody is using
it, there doesn't seem to be a reason to keep backwards compatibility
here.