Add golang to .tool-versions to compile ubi cli
Otherwise, this fails if one has the asdf golang plugin installed, with message like the

    $ bundle exec rake ubi
    cd cli && go build -tags osusergo,netgo
    No version is set for command go
    Consider adding one of the following versions in your config file at /home/fdr/code/clover/.tool-versions
    golang 1.22.5
    golang 1.23.2
    golang 1.23.3
    golang 1.23.4
    golang 1.24.0
    rake aborted!
    Command failed with status (126): [cd cli && go build -tags osusergo,netgo]
    /home/fdr/code/clover/Rakefile:283:in `block in <top (required)>'
    /home/fdr/.asdf/installs/ruby/3.2.7/bin/bundle:25:in `load'
    /home/fdr/.asdf/installs/ruby/3.2.7/bin/bundle:25:in `<main>'
    Tasks: TOP => ubi
    (See full trace by running task with --trace)

An alternative would be to tell .tool-versions to use the "system"
golang if we think forcing people who don't work with the CLI to get
new versions of Go from time to time is too invasive.

On the other hand, bringing golang into asdf's purview means it's very
easy for everyone working with the project to instrument or alter the
cli.