Remove /failover endpoint for Postgres
We initially added this endpoint to allow partners to trigger failover on the
instances they manage, so that they can update the image by refreshing the boot
image. However, it is not being used at the moment, and we would probably do
in-place updates if necessary. For the sake of simplification, we are removing
this endpoint.