Make ProvisionKubernetesNode  cancel itself if cluster is dropped
This patch adds a `before_run` check to ProvisionKubernetesNode prog
so that it drops itself if the cluster is destroyed. The cleanup of the
vm is handled by cluster and nodepool nexus progs.

The nexuses also get a check for waiting for any possible child progs
before continuing with the destroy procedure.