Fix the runner recycling issue for last-second assignments
JIT tokens are valid for 5 minutes. If the runner doesn't pick a job
within that time, the job may have been canceled before assignment, so
we destroy the runner.

Previously, we waited exactly 5 minutes before destroying the runner,
but sometimes GitHub assigns a job at the last second. Even if we check
if the runner is busy with the GitHub API, it can still receive jobs
afterward.

The runner was getting stuck in the destroy because we also checked if
it was busy with the GitHub API during the destruction process.

With this PR, we wait a few extra seconds beyond the 5-minute mark. This
will help avoid the runner recycling issue for last-second assignments.