Avoid Sequel::NoExistingObject exception in race condition in github cache route
This appears to be the most common unhandled exception in production.

The returned row may be deleted between when it was retrieved and
when it is updated later in the route.  While it is possible avoid it
with a transaction and using FOR UPDATE, it doesn't seem necessary.
We can implicitly ignore cases where the row was already deleted,
by switching from Model#update to Dataset#update. Dataset#update
will return 0 and not raise an exception if the model instance has
been destroyed.

This doesn't affect the return value of the route, since that only
uses data stored in the instance, and does not make new database
calls.