Disconnect Sequel::Database instance before fork
Without this, we are getting exceptions in production shortly after
deployment due to multiple processes sharing database connections.

This adds a puma configuration file, and moves the configuration to
that file.

Update bin/production_check to use a clustered, preloaded puma,
since I assume that is what we must be running in production.
Additionally, if the app boots, send a request that requires
database access, and check for the expected response, before
considering it working.